requirements.txt
Flask
datetime 
uuid
Flask-SQLAlchemy
PyJWT
gpt4all
waitress
pathlib